requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
sqlalchemy
pydantic
google-auth
itsdangerous